access right functionality

+++ base/views.py
@@ -78,6 +78,10 @@ def create_room(request):
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.user:
+        return HttpResponse("You are not allowed here!!")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -90,6 +94,8 @@ def update_room(request, pk):
 @login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.user:
+        return HttpResponse("You are not allowed here!!")
     if request.method == "POST":
         room.delete()
         return redirect("home")
